// EscalafonFrontend/src/app/interfaces/v-vinculacion-laboral.ts
export interface VVinculacionLaboral {
    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { string } - cVincLabNumeroDocumento
     */
    cVincLabNumeroDocumento?: string;

    /**
     * @type { Date } - dtVincLabFechaDocumento
     */
    dtVincLabFechaDocumento?: Date;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { number } - iRegLabId
     */
    iRegLabId?: number;

    /**
     * @type { number } - iAccVincId
     */
    iAccVincId?: number;

    /**
     * @type { number } - iMotAccVincId
     */
    iMotAccVincId?: number;

    /**
     * @type { number } - iCentLabId
     */
    iCentLabId?: number;

    /**
     * @type { string } - cVincLabCodigoPlaza
     */
    cVincLabCodigoPlaza?: string;

    /**
     * @type { string } - cVincLabcargo
     */
    cVincLabcargo?: string;

    /**
     * @type { number } - iCatRemuId
     */
    iCatRemuId?: number;

    /**
     * @type { number } - iGrupOcupId
     */
    iGrupOcupId?: number;

    /**
     * @type { number } - iJorLabId
     */
    iJorLabId?: number;

    /**
     * @type { number } - iCargLabId
     */
    iCargLabId?: number;

    /**
     * @type { number } - iVincLabHorasJornadaLaboral
     */
    iVincLabHorasJornadaLaboral?: number;

    /**
     * @type { Date } - dtVincLabFechaInicio
     */
    dtVincLabFechaInicio?: Date;

    /**
     * @type { boolean } - bVincLabMandatoJudicial
     */
    bVincLabMandatoJudicial?: boolean;

    /**
     * @type { Date } - dtVincLabFechaFin
     */
    dtVincLabFechaFin?: Date;

    /**
     * @type { string } - cTipoDocNombre
     */
    cTipoDocNombre?: string;

    /**
     * @type { string } - cRegLabNombre
     */
    cRegLabNombre?: string;

    /**
     * @type { string } - cAccVincNombre
     */
    cAccVincNombre?: string;

    /**
     * @type { string } - cMotAccVincNombre
     */
    cMotAccVincNombre?: string;

    /**
     * @type { string } - cCentLabNombre
     */
    cCentLabNombre?: string;

    /**
     * @type { string } - cCargLabNombre
     */
    cCargLabNombre?: string;

    /**
     * @type { number } - iVincLabId
     */
    iVincLabId?: number;

    /**
     * @type { number } - iEscCatId
     */
    iEscCatId?: number;

    /**
     * @type { string } - cVincLabUseZonaSubRegion
     */
    cVincLabUseZonaSubRegion?: string;

    /**
     * @type { string } - cVincLabAnotaciones
     */
    cVincLabAnotaciones?: string;

}