// EscalafonFrontend/src/app/interfaces/vinculacion-laboral.ts
export interface VinculacionLaboral {
    /**
     * @type { number } - iVincLabId
     */
    iVincLabId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { string } - cVincLabNumeroDocumento
     */
    cVincLabNumeroDocumento?: string;

    /**
     * @type { Date } - dtVincLabFechaDocumento
     */
    dtVincLabFechaDocumento?: Date;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { number } - iRegLabId
     */
    iRegLabId?: number;

    /**
     * @type { number } - iAccVincId
     */
    iAccVincId?: number;

    /**
     * @type { number } - iMotAccVincId
     */
    iMotAccVincId?: number;

    /**
     * @type { number } - iCentLabId
     */
    iCentLabId?: number;

    /**
     * @type { string } - cVincLabCodigoPlaza
     */
    cVincLabCodigoPlaza?: string;

    /**
     * @type { number } - iEscCatId
     */
    iEscCatId?: number;

    /**
     * @type { string } - cVincLabUseZonaSubRegion
     */
    cVincLabUseZonaSubRegion?: string;

    /**
     * @type { string } - cVincLabcargo
     */
    cVincLabcargo?: string;

    /**
     * @type { number } - iCatRemuId
     */
    iCatRemuId?: number;

    /**
     * @type { number } - iGrupOcupId
     */
    iGrupOcupId?: number;

    /**
     * @type { number } - iCargLabId
     */
    iCargLabId?: number;

    /**
     * @type { number } - iJorLabId
     */
    iJorLabId?: number;

    /**
     * @type { number } - iVincLabHorasJornadaLaboral
     */
    iVincLabHorasJornadaLaboral?: number;

    /**
     * @type { Date } - dtVincLabFechaInicio
     */
    dtVincLabFechaInicio?: Date;

    /**
     * @type { Date } - dtVincLabFechaFin
     */
    dtVincLabFechaFin?: Date;

    /**
     * @type { boolean } - bVincLabMandatoJudicial
     */
    bVincLabMandatoJudicial?: boolean;

    /**
     * @type { string } - cVincLabAnotaciones
     */
    cVincLabAnotaciones?: string;

}