// EscalafonFrontend/src/app/interfaces/motivo-accion-vinculacion.ts
export interface MotivoAccionVinculacion {
    /**
     * @type { number } - iMotAccVincId
     */
    iMotAccVincId?: number;

    /**
     * @type { number } - iAccVincId
     */
    iAccVincId?: number;

    /**
     * @type { string } - cMotAccVincNombre
     */
    cMotAccVincNombre?: string;

}